draft of first part of rewrite - seperated functions into classes in files, moving suff around. Did board.js mostly
