+++ index.html
@@ -6,7 +6,7 @@
     <link rel="icon" type="image/x-icon" href="./assets/titanicon.ico">
     <title>teti</title>
     <link rel="stylesheet" href="./style.css">
-    <script defer src="./code.js"></script>
+    <script defer src="./src/main.js" type="module"></script>
 </head>
 <body id="body">
     <div id="board">
