+++ src/pieces.json
@@ -1,0 +1,47 @@
+[{
+    "name": "z",
+    "shape1": [[1, 1, 0], [0, 1, 1], [0, 0, 0]],
+    "shape2": [[0, 0, 1], [0, 1, 1], [0, 1, 0]],
+    "shape3": [[0, 0, 0], [1, 1, 0], [0, 1, 1]],
+    "shape4": [[0, 1, 0], [1, 1, 0], [1, 0, 0]],
+    "colour": "#D83A28"
+}, {
+    "name": "s",
+    "shape1": [[0, 1, 1], [1, 1, 0], [0, 0, 0]],
+    "shape2": [[0, 1, 0], [0, 1, 1], [0, 0, 1]],
+    "shape3": [[0, 0, 0], [0, 1, 1], [1, 1, 0]],
+    "shape4": [[1, 0, 0], [1, 1, 0], [0, 1, 0]],
+    "colour": "#7ACD44"
+}, {
+    "name": "o",
+    "shape1": [[1, 1], [1, 1]],
+    "colour": "#F2D74C"
+}, {
+    "name": "t",
+    "shape1": [[0, 1, 0], [1, 1, 1], [0, 0, 0]],
+    "shape2": [[0, 1, 0], [0, 1, 1], [0, 1, 0]],
+    "shape3": [[0, 0, 0], [1, 1, 1], [0, 1, 0]],
+    "shape4": [[0, 1, 0], [1, 1, 0], [0, 1, 0]],
+    "colour": "#C132D0"
+}, {
+    "name": "j",
+    "shape1": [[1, 0, 0], [1, 1, 1], [0, 0, 0]],
+    "shape2": [[0, 1, 1], [0, 1, 0], [0, 1, 0]],
+    "shape3": [[0, 0, 0], [1, 1, 1], [0, 0, 1]],
+    "shape4": [[0, 1, 0], [0, 1, 0], [1, 1, 0]],
+    "colour": "#3358DD"
+}, {
+    "name": "l",
+    "shape1": [[0, 0, 1], [1, 1, 1], [0, 0, 0]],
+    "shape2": [[0, 1, 0], [0, 1, 0], [0, 1, 1]],
+    "shape3": [[0, 0, 0], [1, 1, 1], [1, 0, 0]],
+    "shape4": [[1, 1, 0], [0, 1, 0], [0, 1, 0]],
+    "colour": "#EDA93F"
+}, {
+    "name": "i",
+    "shape1": [[0, 0, 0, 0], [1, 1, 1, 1], [0, 0, 0, 0], [0, 0, 0, 0]],
+    "shape2": [[0, 0, 1, 0], [0, 0, 1, 0], [0, 0, 1, 0], [0, 0, 1, 0]],
+    "shape3": [[0, 0, 0, 0], [0, 0, 0, 0], [1, 1, 1, 1], [0, 0, 0, 0]],
+    "shape4": [[0, 1, 0, 0], [0, 1, 0, 0], [0, 1, 0, 0], [0, 1, 0, 0]],
+    "colour": "#65DBC8"
+}]
